+++ src/extension/Cycode.VisualStudio.Extension.Shared/Constants.cs
@@ -2,11 +2,12 @@ using System.IO;
 
 namespace Cycode.VisualStudio.Extension.Shared;
 
-public static class Constants
-{
+public static class Constants {
+    public const string AppName = "visual_studio_extension";
+
     public static readonly string PluginPath = GetPluginsPath();
     public static readonly string DefaultCliPath = GetDefaultCliPath();
-    public const string RequiredCliVersion = "1.10.2";
+    public const string RequiredCliVersion = "1.10.3";
 
     public const string CliGithubOrg = "cycodehq";
     public const string CliGithubRepo = "cycode-cli";
@@ -18,8 +19,14 @@ public static class Constants
     public const int PluginAutoSaveFlushInitialDelaySec = 0;
     public const int PluginAutoSaveFlushDelaySec = 5;
 
-    private static string GetPluginsPath()
-    {
+    public const string SentryDsn =
+        "https://091cdc01001e4600a30ac02f1b82c4c5@o1026942.ingest.us.sentry.io/4507543901700096";
+    public const bool SentryDebug = false;
+    public const float SentrySampleRate = 1.0f;
+    public const bool SentrySendDefaultPii = false;
+    public const bool SentryAutoSessionTracking = true;
+
+    private static string GetPluginsPath() {
         string appDataPath = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
         return Path.Combine(appDataPath, "Cycode", "VisualStudioExtension");
     }