+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 
+## [1.1.2] - 2024-07-15
+
+- Integrate Sentry
+- Fix extension loading when the core is not found
+
 ## [1.1.1] - 2024-07-05
 
 - Fix extension loading on the background thread
@@ -20,6 +25,8 @@
 
 The first public release of the extension.
 
+[1.1.2]: https://github.com/cycodehq/visual-studio-extension/releases/tag/v1.1.2
+
 [1.1.1]: https://github.com/cycodehq/visual-studio-extension/releases/tag/v1.1.1
 
 [1.1.0]: https://github.com/cycodehq/visual-studio-extension/releases/tag/v1.1.0
@@ -28,4 +35,4 @@ The first public release of the extension.
 
 [1.0.0]: https://github.com/cycodehq/visual-studio-extension/releases/tag/v1.0.0
 
-[Unreleased]: https://github.com/cycodehq/visual-studio-extension/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/cycodehq/visual-studio-extension/compare/v1.1.2...HEAD